ver up

+++ src/background.js
@@ -1,14 +1,51 @@
-let setAuthToken = () => {
-	chrome.identity.getAuthToken({ interactive: true }, (token) => {
-		chrome.storage.local.set({'token': token});
-	});
-};
-
 let createContextMenus = (param) => {
 	chrome.contextMenus.create(Object.assign({
 		parentId: 'CalendarGroupTop',
 		documentUrlPatterns: ['https://calendar.google.com/calendar/*']
 	}, param));
+};
+
+let createPromiseCallback = (result, reject) => {
+	return (arg) => {
+		chrome.runtime.lastError ? result(arg) : reject(chrome.runtime.lastError);
+	};
+}
+
+let getAuthToken = (param) => {
+	return new Promise((result, reject) => {
+		chrome.identity.getAuthToken({ interactive: true }, createPromiseCallback(result, reject));
+	}).catch((err) => console.log(err));
+};
+
+let setAllContextMenus = () => {
+	chrome.contextMenus.removeAll(() => {
+		createContextMenus({
+			id: 'CalendarGroupTop',
+			title: 'Calendar Group',
+			parentId: undefined
+		});
+		chrome.storage.local.get('calendarGroup', (value) => {
+			let calendarGroup = value.calendarGroup;
+			(calendarGroup || []).forEach((group) => {
+				createContextMenus({
+					id: 'CalendarGroup:' + group.ids,
+					title: group.name
+				});
+			});
+			createContextMenus({
+				id: 'CalendarGroupCreate',
+				title: 'add current group'
+			});
+			createContextMenus({
+				id: 'ClearCalendarGroup',
+				title: 'clear all group'
+			});
+			createContextMenus({
+				id: 'ClearAuthToken',
+				title: 'clear auth token'
+			});
+		});
+	});
 };
 
 let timeout;
@@ -25,37 +62,10 @@ function refreshContextMenus(tabId) {
 			if (!tab.url.match(/^https:\/\/calendar\.google\.com\//)) {
 				return;
 			}
-			setAuthToken();
-			chrome.contextMenus.removeAll(() => {
-				createContextMenus({
-					id: 'CalendarGroupTop',
-					title: 'Calendar Group',
-					parentId: undefined
-				});
-				chrome.storage.local.get('calendarGroup', (value) => {
-					let calendarGroup = value.calendarGroup;
-					(calendarGroup || []).forEach((group) => {
-						createContextMenus({
-							id: 'CalendarGroup:' + group.ids,
-							title: group.name
-						});
-					});
-					createContextMenus({
-						id: 'CalendarGroupCreate',
-						title: 'add current group'
-					});
-					createContextMenus({
-						id: 'ClearCalendarGroup',
-						title: 'clear all group'
-					});
-				});
-			});
+			setAllContextMenus();
 		});
 	}, 3000);
 }
-
-chrome.runtime.onInstalled.addListener(setAuthToken);
-chrome.tabs.onUpdated.addListener(refreshContextMenus);
 
 let fetchApi = (url, param) => {
 	let baseUrl = 'https://www.googleapis.com/calendar/v3/users/me/calendarList';
@@ -63,17 +73,26 @@ let fetchApi = (url, param) => {
 		'headers': {}
 	};
 	return new Promise((resolve, reject) => {
-		chrome.storage.local.get('token', (value) => {
+		getAuthToken().then((token) => {
 			Object.assign(param.headers, {
-				'Authorization': 'Bearer ' + value.token
+				'Authorization': 'Bearer ' + token
 			});
-			fetch(baseUrl + url, param).then(resolve, reject);
+			fetch(`${baseUrl}{url}`, param).then(resolve, reject);
 		});
 	});
 };
 
 let fetchSelectedCalendar = () => {
-	return fetchApi('?maxResults=250&fields=items(backgroundColor%2CcolorId%2Cdeleted%2CforegroundColor%2Cid%2Cselected)').then((res) => res.json());
+	let fields = [
+		'summary',
+		'backgroundColor',
+		'colorId',
+		'deleted',
+		'foregroundColor',
+		'id',
+		'selected',
+	].join('%2C');
+	return fetchApi(`?maxResults=250&fields=items(${fields})`).then((res) => res.json());
 }
 
 function createGroup (clickData) {
@@ -84,14 +103,16 @@ function createGroup (clickData) {
 			.map((item) => item.id).join(':')
 		;
 		createContextMenus({
-			id: 'CalendarGroup:' + ids,
+			id: `CalendarGroup:${ids}`,
 			title: name
 		});
 		chrome.storage.local.get('calendarGroup', (value) => {
 			chrome.storage.local.set({'calendarGroup': (value.calendarGroup || []).concat({
 				name: name,
 				ids: ids,
-			})});
+			})}, () => {
+				setAllContextMenus();
+			});
 		});
 	});
 }
@@ -99,7 +120,7 @@ function createGroup (clickData) {
 function getBodyParam (selected, item) {
 	delete item['id'];
 	return {
-		body: JSON.stringify(Object.assign(item, {
+		body: JSON.stringify(Object.assign({}, item, {
 			'selected': selected
 		}))
 	};
@@ -112,10 +133,10 @@ function selectGroup (menuItemId) {
 		let changeCalendarState = (state) => {
 			return items
 				.filter((item) => state === ids.includes(item.id))
-				.filter((item) => state !== item.selected)
+				.filter((item) => state != item.selected)
 				.map((item) => {
 					let id = encodeURIComponent(item.id);
-					let p = Object.assign(param, getBodyParam(state, item));
+					let p = Object.assign({}, param, getBodyParam(state, item));
 					return fetchApi(`/${id}`, p);
 				})
 			;
@@ -141,6 +162,9 @@ chrome.contextMenus.onClicked.addListener((clickData) => {
 		return;
 	}
 	if (clickData.menuItemId === 'ClearCalendarGroup') {
+		if (!confirm('clear all group?')) {
+			return;
+		}
 		chrome.storage.local.clear(() => {
 			chrome.contextMenus.removeAll(() => {
 				chrome.tabs.reload();
@@ -148,8 +172,23 @@ chrome.contextMenus.onClicked.addListener((clickData) => {
 		});
 		return;
 	}
+	if (clickData.menuItemId.match(/^ClearAuthToken/)) {
+		getAuthToken().then((token) => {
+			let promises = [
+				(new Promise((resolve, reject) => {
+					chrome.identity.removeCachedAuthToken({ token }, createPromiseCallback(resolve, reject));
+				})),
+				fetch(`https://accounts.google.com/o/oauth2/revoke?token=${token}`)
+			];
+			return Promise.all(promises);
+		}).then(getAuthToken);
+		return;
+	}
 	if (clickData.menuItemId.match(/^CalendarGroup:/)) {
-		selectGroup(clickData.menuItemId)
+		selectGroup(clickData.menuItemId);
 		return;
 	}
 });
+
+chrome.runtime.onInstalled.addListener(getAuthToken);
+chrome.tabs.onUpdated.addListener(refreshContextMenus);
